+++ src/Main.java
@@ -2,8 +2,8 @@ public class Main {
     public static void main(String[] args) {
         BmiService calculate = new BmiService();
         double height = 1.87; // meters
-        long weight = 98; // kilograms
-        long massIndex = calculate.bodyMassIndex(height, weight);
+        int weight = 98; // kilograms
+        int massIndex = calculate.bodyMassIndex(height, weight);
         System.out.println(massIndex);
     }
 }
